profile: call DiveCalculatedTissue::setVisible() correctly.
This one is rich: when changing to profile-mode, the calculated
tissues are set according to the prefs.calcalltissues flag.

The DiveCalculatedTissue::setVisible() function ignores the
parameter and insteads sets the visibility according to the
expression "prefs.calcalltissues && prefs.calcceiling".

This is because the function is also called by signals,
which provide the wrong parameter.

Pass the correct parameter in the first place. Remove the
crazy signals later.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>